M?J

+++ Web/app/Controller/UsersController.php
@@ -47,8 +47,18 @@ class UsersController extends AppController
 		if($this->request->is('post')){
 			$t = $this->request->data;
 			if(!isset($t['User']['birthdate'])) $t['User']['birthdate'] = 0;
+			if(!isset($t['User']['firstname'])) $t['User']['firstname'] = '';
+			if(!isset($t['User']['lastname'])) $t['User']['lastname'] = '';
+			if(!isset($t['User']['gender'])) $t['User']['gender'] = '';
+			if(!isset($t['User']['email'])) $t['User']['email'] = '';
+			if(!isset($t['User']['phone'])) $t['User']['phone'] = '';
+			if(!isset($t['User']['zipcode'])) $t['User']['zipcode'] = '';
+			if(!isset($t['User']['street'])) $t['User']['street'] = '';
+			if(!isset($t['User']['town'])) $t['User']['town'] = '';
+			if(!isset($t['User']['country'])) $t['User']['country'] = '';
+			if(!isset($t['User']['role'])) $t['User']['role'] = '';
 			if($t['User']['birthdate'] == null) $t['User']['birthdate'] = 0;
-			if($this->User->save($t, true)){
+			if($this->User->save($t)){
 				$this->Session->setFlash('Votre compte a bien été crée.', "message", array('type' => 'success'));
 				$this->redirect(array('controller' => 'cocktails', 'action' => 'index'));
 			}
@@ -69,6 +79,7 @@ class UsersController extends AppController
 		}
 			$this->User->id = $id;
 			$t = $this->request->data;
+			if($t['User']['birthdate'] == null) $t['User']['birthdate'] = 0;
 			if($this->User->save($t, true, array('password','firstname','lastname','birthdate', 'street' ,'town','gender','zipcode','country','phone', 'email'))){
 				$this->Session->setFlash('Votre profil a correctement été mis à jour.','message',array('type' => 'success'));
 				$this->redirect('/');
@@ -129,11 +140,25 @@ class UsersController extends AppController
 	*/
 	public function login() 
 	{
-		$this->Session->delete('Cart');
 		if(!empty($this->data)){
 			if($this->Auth->login()){
 				$this->Session->setFlash('Bienvenue '. $this->Auth->user('firstname') . ' ' . $this->Auth->user('username') .', vous êtes désormais connecté.', 'message', array('type' => 'success'));
+				$idUser = $this->Session->read('Auth.User.id');
+				$this->loadModel('Cart');
+
+				if($this->Session->read('Cart') != null)
+				{
+					foreach ($this->Session->read('Cart.recipe') as $k) {
+						if($k == null) continue;
+						else{
+							$this->request->data['Cart']['recipe'] = $k;
+							$this->request->data['Cart']['user'] = $idUser;
+							$this->Cart->save($this->request->data);
+						}
+					}
+					$this->Session->delete('Cart');
 					$this->redirect(array('controller' => 'recipes', 'action' => 'index'));
+				}
 			}
 			else{
 				$this->Session->setFlash('Veuillez vérifier vos identifiants.', 'message', array('type' => 'danger'));
